+++ engine/flink/management/periodic/src/main/resources/db/batch_periodic/migration/common/V1_107__add_retry_at_and_retries_to_periodic_process_deployments.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "periodic_process_deployments" ADD COLUMN "retries_left" INTEGER DEFAULT 0 NOT NULL;
+ALTER TABLE "periodic_process_deployments" ADD COLUMN "next_retry_at" TIMESTAMP DEFAULT NULL;