+++ engine/flink/management/periodic/src/test/scala/pl/touk/nussknacker/engine/management/periodic/PeriodicDeploymentManagerTest.scala
@@ -15,6 +15,7 @@ import pl.touk.nussknacker.test.PatientScalaFutures
 
 import java.time.Clock
 import scala.concurrent.Await
+import scala.concurrent.duration.Duration
 
 class PeriodicDeploymentManagerTest extends FunSuite
   with Matchers
@@ -41,6 +42,7 @@ class PeriodicDeploymentManagerTest extends FunSuite
       scheduledProcessesRepository = repository,
       EmptyListener,
       DefaultAdditionalDeploymentDataProvider,
+      DeploymentRetryConfig(),
       ProcessConfigEnricher.identity,
       Clock.systemDefaultZone()
     )